server/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
attrs==25.3.0
bcrypt==4.3.0
bidict==0.23.1
blinker==1.9.0
cffi==1.17.1
click==8.1.8
cryptography==44.0.2
Flask==3.1.0
Flask-SocketIO==5.5.1
frozenlist==1.5.0
gunicorn==23.0.0
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jeepney==0.9.0
Jinja2==3.1.6
keyring==25.6.0
MarkupSafe==3.0.2
more-itertools==10.6.0
multidict==6.2.0
packaging==24.2
propcache==0.3.1
psutil==7.0.0
pycparser==2.22
PyJWT==2.10.1
python-engineio==4.11.2
python-socketio==5.12.1
SecretStorage==3.3.3
simple-websocket==1.1.0
websockets==14.2
Werkzeug==3.1.3
wsproto==1.2.0
yarl==1.18.3
